requirements.txt
streamlit
google-generativeai
youtube-transcript-api
yt-dlp
jinja2